+++ demo/lib/examples/document.dart
@@ -256,22 +256,25 @@ Future<Uint8List> generateDocument(
             pw.Paragraph(
                 text:
                     'The PDF file format has changed several times, and continues to evolve, along with the release of new versions of Adobe Acrobat. There have been nine versions of PDF and the corresponding version of the software:'),
-            pw.Table.fromTextArray(context: context, data: const <List<String>>[
-              <String>['Date', 'PDF Version', 'Acrobat Version'],
-              <String>['1993', 'PDF 1.0', 'Acrobat 1'],
-              <String>['1994', 'PDF 1.1', 'Acrobat 2'],
-              <String>['1996', 'PDF 1.2', 'Acrobat 3'],
-              <String>['1999', 'PDF 1.3', 'Acrobat 4'],
-              <String>['2001', 'PDF 1.4', 'Acrobat 5'],
-              <String>['2003', 'PDF 1.5', 'Acrobat 6'],
-              <String>['2005', 'PDF 1.6', 'Acrobat 7'],
-              <String>['2006', 'PDF 1.7', 'Acrobat 8'],
-              <String>['2008', 'PDF 1.7', 'Acrobat 9'],
-              <String>['2009', 'PDF 1.7', 'Acrobat 9.1'],
-              <String>['2010', 'PDF 1.7', 'Acrobat X'],
-              <String>['2012', 'PDF 1.7', 'Acrobat XI'],
-              <String>['2017', 'PDF 2.0', 'Acrobat DC'],
-            ]),
+            pw.TableHelper.fromTextArray(
+              context: context,
+              data: const <List<String>>[
+                <String>['Date', 'PDF Version', 'Acrobat Version'],
+                <String>['1993', 'PDF 1.0', 'Acrobat 1'],
+                <String>['1994', 'PDF 1.1', 'Acrobat 2'],
+                <String>['1996', 'PDF 1.2', 'Acrobat 3'],
+                <String>['1999', 'PDF 1.3', 'Acrobat 4'],
+                <String>['2001', 'PDF 1.4', 'Acrobat 5'],
+                <String>['2003', 'PDF 1.5', 'Acrobat 6'],
+                <String>['2005', 'PDF 1.6', 'Acrobat 7'],
+                <String>['2006', 'PDF 1.7', 'Acrobat 8'],
+                <String>['2008', 'PDF 1.7', 'Acrobat 9'],
+                <String>['2009', 'PDF 1.7', 'Acrobat 9.1'],
+                <String>['2010', 'PDF 1.7', 'Acrobat X'],
+                <String>['2012', 'PDF 1.7', 'Acrobat XI'],
+                <String>['2017', 'PDF 2.0', 'Acrobat DC'],
+              ],
+            ),
             pw.Padding(padding: const pw.EdgeInsets.all(10)),
             pw.Paragraph(
                 text: